Add Default File Logging
A default file log, or log, is now created on usage of the library. A log is created in the `temp` folder of the user, typically `C:\Users\username\AppData\Local\Temp` on a Windows 7 Operating System (OS), if file is not configured differently. The root folder can be changed to any folder on disk. A folder is created in the root folder based on the configured `Application Name`. The default `Application Name` is "LV-Logger". This is known as the "base" folder. Within the `base` folder, a calendar folder structure, or tree, is created. The following format is used: `..\..\YYYY\mm\dd`, where `YYYY` is the year with century, `mm` is the month with leading zero, and `dd` is the day of the month with leading zero. A log file is created and placed in the appropriate calendar folder tree based on the start time/date of the log, or the time and date the log is created. The log is named following the ISO8601 time and date format, i.e. `YYYYmmddTHHMMSS`, where `YYYY`, `mm`, and `dd` are the same as before and `HH` is the hour in 24-hour format, `MM` is the minutes of the hour with leading zero, and `SS` is the seconds of the minute with leading zero. The calendar folder tree and log file name are used regardless of the configuration and currently cannot be changed. This is to prevent logs from being overwritten and easy to find based on time and date.

Log messages are saved as Comma-Separated Values (CSV) strings in the log file. Thus, the log file has the `.csv` file extension.

The file logging system can be configured with a "File Size Limit", which determines when a new log file should be created, or rolled over. The value is in bytes and the default is set to 15MB(ytes). The default is a good size to allow sharing of the log files via email and not be too cumbersome to open and review. The default can be changed with the related configure Virtual Instruments (VIs) provided in the library.

git-svn-id: file:///D/Users/cfield/Repos/Logger/trunk@24 e2ff8c16-26f6-4d46-883d-020078979db7