Add ability to ignore multiple errors
The `Ignore Errors.vi` of the Public API allows for multiple errors to be ignored. An array of error codes is passed to the VI and any errors matching the error codes in the array are converted to "Benign" errors. They are still logged by the verbosity must be increased to `4` to have a dialog appear. Bengin errors are logged with the DEBUG tag since it is useful during debugging to know all errors that are occurring in an application.

git-svn-id: file:///D/Users/cfield/Repos/Logger/trunk@63 e2ff8c16-26f6-4d46-883d-020078979db7