Add utility VI to create a folder log root folder in the user's temp directory
The `Create Temporary Log Folder.vi` takes an array of strings with each string a folder name where index = 0 is the top level folder, and appends the folder structure to the user's temporary folder. The `Log` folder is appended to the end of the path. This can be used to create the root folder for the file log, where the calendar folder structure is created automatically within this temporary root folder.

git-svn-id: svn://home-office.fios-router.home/Logger/trunk@102 e2ff8c16-26f6-4d46-883d-020078979db7