Change CSV string creation to use a buffer
A buffer (an array of U8 integers) is allocated the first time the `Create Bulk CSV String` is called. A control is used to set the capacity of the buffer. The buffer is converted to a string and the `Replace Subset` primitive is instead of the `String Concatenate` primitive to create the CSV string that is written to the file. This eliminates creating (allocating) a string each time statements are written to the logging file. This should improve performance and reduce memory fragmentation on real-time systems.

git-svn-id: svn://home-office.fios-router.home/Logger/branches/fix-reduce-dynamic-allocations@170 e2ff8c16-26f6-4d46-883d-020078979db7