Fix the Show in Palette button for the VI Package Manager not working correctly
Ultimately, the fix was to change the VI title of the Initialize VI from Initialize.vi to Initialize within the palette editor of the VI Package configuration dialog. Apparently, if there is a VI named "Initialize.vi" in the palette, then the VI Package Manager will point to the Agilent 34401 library. This appears to be a bug in the VI Package Manager, but changing the title is a suitable workaround.

The version numbers and dates were updated in the package's readme as well.

git-svn-id: svn://home-office.fios-router.home/Logger/trunk@247 e2ff8c16-26f6-4d46-883d-020078979db7