Remove LabVIEW 2009 support
LabVIEW 2009 does not support inlining sub-VIs. This causes problems with the Call Chain function and Source field in logging statements. The simplest solution is to drop support for LabVIEW 2009. I am not aware of anyone using Logger with LabVIEW 2009, but if they are and would like an upgrade, then they can just special request it from me.

git-svn-id: file:///D/Users/cfield/Repos/Logger/trunk@318 e2ff8c16-26f6-4d46-883d-020078979db7